+++ src/components/cart/Cart.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script>
-import CartItem from '@/components/checkout/CartItem.vue';
+import CartItem from '@/components/cart/CartItem.vue';
 import { mapGetters } from 'vuex';
 
 export default {
