+++ src/components/cart/EmptyCart.vue
@@ -1,0 +1,24 @@
+<template>
+  <div
+    class="rounded-lg max-w-xs p-8 mb-8 m-auto border border-dashed border-white text-center"
+  >
+    <img
+      src="https://res.cloudinary.com/whynotearth/image/upload/v1586099571/foodouken/heart-1x_axcey4.png"
+      alt="Empty cart"
+      class="m-auto"
+    />
+    <p class="mt-6 text-gray-500 font-bold text-2xl">Nothing to eat here!</p>
+    <p class="mt-2 text-gray-500 text-lg">
+      You should grab your
+      <br />favorites!
+    </p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'EmptyCart'
+};
+</script>
+
+<style></style>
