+++ src/views/Home.vue
@@ -13,12 +13,8 @@
         <card-holder />
       </div>
       <div class="lg:w-4/12 lg:ml-4">
-        <h3 class="text-5xl text-white font-bold text-center mb-8">Cart</h3>
         <cart v-if="cart.length" />
         <empty-cart v-else />
-        <p class="text-gray-500 text-center text-lg">Estimated Delivery Time: 45 minutes.</p>
-        <div class="w-full text-center my-4" v-if="cart.length">
-          <Button title="Order now" />
         <h3
           v-if="!renderForm" 
           class="text-5xl text-white font-bold text-center mb-8"
@@ -26,11 +22,12 @@
           Cart
         </h3>
         <checkout-form v-if="renderForm" />
-        <cart />
+        <cart v-if="cart.length" />
+        <empty-cart v-else />
         <p class="text-gray-500 text-center text-lg">
           Estimated Delivery Time: 45 minutes.
         </p>
-        <div v-if="!renderForm" class="w-full text-center my-4">
+        <div v-if="cart.length || !renderForm" class="w-full text-center my-4">
           <Button title="Order now" @clicked="renderForm = true" />
         </div>
       </div>
@@ -43,8 +40,8 @@ import HeroSection from '@/components/HeroSection.vue';
 import CardHolder from '@/components/cards/CardHolder.vue';
 import CategoryHolder from '@/components/categories/CategoryHolder.vue';
 import Button from '@/components/Button.vue';
-import Cart from '@/components/checkout/Cart.vue';
-import EmptyCart from '@/components/checkout/EmptyCart.vue';
+import Cart from '@/components/cart/Cart.vue';
+import EmptyCart from '@/components/cart/EmptyCart.vue';
 import { mapGetters, mapActions } from 'vuex';
 
 export default {
@@ -55,7 +52,7 @@ export default {
     CategoryHolder,
     Button,
     Cart,
-    EmptyCart
+    EmptyCart,
     CheckoutForm: () => import('@/components/forms/CheckoutForm.vue')
   },
   data() {
